Unit test for api and dusk test for installer

+++ app/Http/Controllers/Api/v1/TicketController.php
@@ -30,15 +30,14 @@ use Mail;
  *
  * @version v1
  */
-class TicketController extends Controller
-{
+class TicketController extends Controller {
+
     /**
      * Create a new controller instance.
      *
      * @return type response
      */
-    public function __construct()
-    {
+    public function __construct() {
         $PhpMailController = new PhpMailController();
         $this->PhpMailController = $PhpMailController;
     }
@@ -55,8 +54,7 @@ class TicketController extends Controller
      *
      * @return type string
      */
-    public function createTicket($user_id, $subject, $body, $helptopic, $sla, $priority, $source, $headers, $dept, $assignto, $form_data, $attach = '')
-    {
+    public function createTicket($user_id, $subject, $body, $helptopic, $sla, $priority, $source, $headers, $dept, $assignto, $form_data, $attach = '') {
         try {
             //return $headers;
             $max_number = Tickets::whereRaw('id = (select max(`id`) from tickets)')->first();
@@ -119,8 +117,7 @@ class TicketController extends Controller
      *
      * @return type
      */
-    public function storeCollaborators($headers, $id)
-    {
+    public function storeCollaborators($headers, $id) {
         try {
             //return $headers;
             $company = $this->company();
@@ -173,8 +170,7 @@ class TicketController extends Controller
      *
      * @return type
      */
-    public function ticketThread($subject, $body, $id, $user_id)
-    {
+    public function ticketThread($subject, $body, $id, $user_id) {
         try {
             $thread = new Ticket_Thread();
             $thread->user_id = $user_id;
@@ -197,8 +193,7 @@ class TicketController extends Controller
      *
      * @return type integer
      */
-    public function ticketNumber($ticket_number)
-    {
+    public function ticketNumber($ticket_number) {
         try {
             //dd($ticket_number);
             $number = $ticket_number;
@@ -238,8 +233,7 @@ class TicketController extends Controller
      *
      * @return type string
      */
-    public function generateRandomString($length = 10)
-    {
+    public function generateRandomString($length = 10) {
         try {
             $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
             $charactersLength = strlen($characters);
@@ -262,8 +256,7 @@ class TicketController extends Controller
      *
      * @return type bool
      */
-    public function reply($thread, $request, $ta, $attach = '')
-    {
+    public function reply($thread, $request, $ta, $attach = '') {
         try {
             $check_attachment = null;
             $eventthread = $thread->where('ticket_id', $request->input('ticket_ID'))->first();
@@ -296,7 +289,7 @@ class TicketController extends Controller
                 $thread2->ticket_id = $thread->ticket_id;
                 $thread2->user_id = Auth::user()->id;
                 $thread2->is_internal = 1;
-                $thread2->body = 'This Ticket have been assigned to '.Auth::user()->first_name.' '.Auth::user()->last_name;
+                $thread2->body = 'This Ticket have been assigned to ' . Auth::user()->first_name . ' ' . Auth::user()->last_name;
                 $thread2->save();
             }
             if ($tickets->status > 1) {
@@ -357,7 +350,7 @@ class TicketController extends Controller
                 if ($user_id_collab->role == 'user') {
                     $collab_user_name = $user_id_collab->user_name;
                 } else {
-                    $collab_user_name = $user_id_collab->first_name.' '.$user_id_collab->last_name;
+                    $collab_user_name = $user_id_collab->first_name . ' ' . $user_id_collab->last_name;
                 }
 //                 Mail::send('emails.ticket_re-reply', ['content' => $reply_content, 'ticket_number' => $ticket_number, 'From' => $company, 'name' => $collab_user_name, 'Agent_Signature' => $agentsign], function ($message) use ($collab_email, $collab_user_name, $ticket_number, $ticket_subject, $check_attachment) {
 //                     $message->to($collab_email, $collab_user_name)->subject($ticket_subject . '[#' . $ticket_number . ']');
@@ -372,6 +365,7 @@ class TicketController extends Controller
                 try {
                     $this->PhpMailController->sendmail($from = $this->PhpMailController->mailfrom('0', $ticketdata->dept_id), $to = ['user' => $admin_user, 'email' => $admin_email], $message = ['subject' => $updated_subject, 'body' => $body, 'scenario' => $mail], $template_variables = ['ticket_agent_name' => $admin_user, 'ticket_client_name' => $username, 'ticket_client_email' => $emailadd, 'user' => $admin_user, 'ticket_number' => $ticket_number2, 'email_address' => $emailadd, 'name' => $ticket_creator]);
                 } catch (\Exception $e) {
+                    
                 }
             }
 
@@ -387,8 +381,7 @@ class TicketController extends Controller
      *
      * @return type
      */
-    public function company()
-    {
+    public function company() {
         try {
             $company = Company::Where('id', '=', '1')->first();
             if ($company->company_name == null) {
@@ -411,8 +404,7 @@ class TicketController extends Controller
      *
      * @return type bool
      */
-    public function ticketEditPost($ticket_id, $thread, $ticket)
-    {
+    public function ticketEditPost($ticket_id, $thread, $ticket) {
         try {
             $ticket = $ticket->where('id', '=', $ticket_id)->first();
 
@@ -440,13 +432,12 @@ class TicketController extends Controller
      *
      * @return type bool
      */
-    public function assign($id)
-    {
+    public function assign($id) {
         try {
             $UserEmail = Input::get('user');
             //dd($UserEmail);
             // $UserEmail = 'sujitprasad12@yahoo.in';
-            $user = User::where('email', '=', $UserEmail)->first();
+            $user = User::where('id', '=', $UserEmail)->first();
             if (!$user) {
                 return ['error' => 'No agent not found'];
             }
@@ -467,7 +458,7 @@ class TicketController extends Controller
             $thread->ticket_id = $ticket->id;
             $thread->user_id = Auth::user()->id;
             $thread->is_internal = 1;
-            $thread->body = 'This Ticket has been assigned to '.$user->first_name.' '.$user->last_name;
+            $thread->body = 'This Ticket has been assigned to ' . $user->first_name . ' ' . $user->last_name;
             $thread->save();
 
             $company = $this->company();
@@ -476,7 +467,7 @@ class TicketController extends Controller
             $agent = $user->first_name;
             $agent_email = $user->email;
 
-            $master = Auth::user()->first_name.' '.Auth::user()->last_name;
+            $master = Auth::user()->first_name . ' ' . Auth::user()->last_name;
             if (Alert::first()->internal_status == 1 || Alert::first()->internal_assigned_agent == 1) {
                 // // ticket assigned send mail
                 // Mail::send('emails.Ticket_assign', ['agent' => $agent, 'ticket_number' => $ticket_number, 'from' => $company, 'master' => $master, 'system' => $system], function ($message) use ($agent_email, $agent, $ticket_number, $ticket_subject) {
@@ -484,7 +475,7 @@ class TicketController extends Controller
                 // });
 
                 try {
-                    $this->PhpMailController->sendmail($from = $this->PhpMailController->mailfrom('0', $ticket->dept_id), $to = ['name' => $agent, 'email' => $agent_email], $message = ['subject' => $ticket_subject.'[#'.$ticket_number.']', 'scenario' => 'assign-ticket'], $template_variables = ['ticket_agent_name' => $agent, 'ticket_number' => $ticket_number, 'ticket_assigner' => $master]);
+                    $this->PhpMailController->sendmail($from = $this->PhpMailController->mailfrom('0', $ticket->dept_id), $to = ['name' => $agent, 'email' => $agent_email], $message = ['subject' => $ticket_subject . '[#' . $ticket_number . ']', 'scenario' => 'assign-ticket'], $template_variables = ['ticket_agent_name' => $agent, 'ticket_number' => $ticket_number, 'ticket_assigner' => $master]);
                 } catch (\Exception $e) {
                     return 0;
                 }
@@ -504,8 +495,7 @@ class TicketController extends Controller
      *
      * @return type string
      */
-    public function delete($ids, $ticket)
-    {
+    public function delete($ids, $ticket) {
         try {
             foreach ($ids as $id) {
                 $ticket_delete = $ticket->where('id', '=', $id)->first();
@@ -537,7 +527,7 @@ class TicketController extends Controller
                         $thread->ticket_id = $ticket_delete->id;
                         $thread->user_id = Auth::user()->id;
                         $thread->is_internal = 1;
-                        $thread->body = $ticket_status_message->message.' '.Auth::user()->first_name.' '.Auth::user()->last_name;
+                        $thread->body = $ticket_status_message->message . ' ' . Auth::user()->first_name . ' ' . Auth::user()->last_name;
                         $thread->save();
                     }
                 } else {
@@ -558,8 +548,7 @@ class TicketController extends Controller
      *
      * @return type bool
      */
-    public function checkEmail($email)
-    {
+    public function checkEmail($email) {
         try {
             $check = User::where('email', '=', $email)->first();
             if ($check) {
@@ -577,8 +566,7 @@ class TicketController extends Controller
      *
      * @return type
      */
-    public function system()
-    {
+    public function system() {
         try {
             $system = System::Where('id', '=', '1')->first();
             if ($system->name == null) {
@@ -601,8 +589,7 @@ class TicketController extends Controller
      *
      * @return int
      */
-    public function attach($thread, $attach)
-    {
+    public function attach($thread, $attach) {
         try {
             $ta = new Ticket_attachments();
             foreach ($attach as $file) {
@@ -621,10 +608,9 @@ class TicketController extends Controller
      *
      * @return type json
      */
-    public function autosearch()
-    {
+    public function autosearch() {
         $term = \Input::get('term');
-        $user = \App\User::where('email', 'LIKE', '%'.$term.'%')->orWhere('first_name', 'LIKE', '%'.$term.'%')->orWhere('last_name', 'LIKE', '%'.$term.'%')->orWhere('user_name', 'LIKE', '%'.$term.'%')->pluck('email');
+        $user = \App\User::where('email', 'LIKE', '%' . $term . '%')->orWhere('first_name', 'LIKE', '%' . $term . '%')->orWhere('last_name', 'LIKE', '%' . $term . '%')->orWhere('user_name', 'LIKE', '%' . $term . '%')->pluck('email');
 
         return $user;
     }
@@ -636,8 +622,7 @@ class TicketController extends Controller
      *
      * @return type json
      */
-    public function useradd()
-    {
+    public function useradd() {
         $email = Input::get('email');
         $ticket_id = Input::get('ticket_id');
         $company = $this->company();
@@ -670,26 +655,25 @@ class TicketController extends Controller
      *
      * @return type
      */
-    public function userremove()
-    {
+    public function userremove() {
         $email = Input::get('email');
         $ticketid = Input::get('ticketid');
         $user = new User();
         $user = $user->where('email', $email)->first();
-        $ticket_collaborator = Ticket_Collaborator::where('ticket_id', '=', $ticketid)
-                ->where('user_id', $user->id)
-                ->first();
-        if ($ticket_collaborator) {
-            $ticket_collaborator->delete();
+        if ($user) {
+            $ticket_collaborator = Ticket_Collaborator::where('ticket_id', '=', $ticketid)
+                    ->where('user_id', $user->id)
+                    ->first();
+            if ($ticket_collaborator) {
+                $ticket_collaborator->delete();
 
-            return 'deleted successfully';
-        } else {
-            return 'not found';
+                return 'deleted successfully';
+            }
         }
+        return 'not found';
     }
 
-    public function getCollaboratorForTicket()
-    {
+    public function getCollaboratorForTicket() {
         try {
             $ticketid = Input::get('ticket_id');
 
@@ -717,13 +701,12 @@ class TicketController extends Controller
         }
     }
 
-    public function avatarUrl($email)
-    {
+    public function avatarUrl($email) {
         try {
             $user = new User();
             $user = $user->where('email', $email)->first();
             if ($user->profile_pic) {
-                $url = url('uploads/profilepic/'.$user->profile_pic);
+                $url = url('uploads/profilepic/' . $user->profile_pic);
             } else {
                 $url = \Gravatar::src($email);
             }
@@ -734,4 +717,5 @@ class TicketController extends Controller
             throw new \Exception($ex->getMessage());
         }
     }
+
 }
